refactore: chamnge files of bot delivery
