+++ internal/delivery/bot/handlers.go
@@ -1,6 +1,24 @@
 package bot
 
-import tele "gopkg.in/telebot.v3"
+import (
+	"sync"
+
+	"github.com/arya237/foodPilot/internal/services/restaurant"
+	tele "gopkg.in/telebot.v3"
+)
+
+type handler struct {
+	restaurant restaurant.Connector
+	mu         sync.RWMutex
+	cache      map[int64]userState
+}
+
+func newHandler(restaurant restaurant.Connector) *handler {
+	return &handler{
+		cache:      make(map[int64]userState),
+		restaurant: restaurant,
+	}
+}
 
 func onStart(c tele.Context) error {
 	keyboard := &tele.ReplyMarkup{
@@ -20,5 +38,3 @@ func onStart(c tele.Context) error {
 func aboutUs(c tele.Context) error {
 	return c.Send("ما خیلی خفنیم")
 }
-
-
